% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as2}
\alias{as2}
\title{A more robust form of the R \code{\link{as}} function.}
\usage{
as2(object, class)
}
\arguments{
\item{object}{any \code{R} object}

\item{class}{the name of the class to which \code{object} should be coerced}
}
\description{
Alternative to \code{as} that allows any data object to be converted to any other.
}

