% Generated by roxygen2 (4.0.0): do not edit by hand
\name{buffer}
\alias{buffer}
\title{Pads an object to a desired length, either with replicates of itself or another repeated object.}
\usage{
buffer(x, length.out = len(x), fill = NULL, preserveClass = TRUE)
}
\arguments{
\item{x}{an R object}

\item{length.out}{the desired length of the final output}

\item{fill}{R object to fill empty rows in columns below the max size.  If unspecified, repeats input rows in the same way as \code{cbind}.}

\item{preserveClass}{determines whether to return an object of the same class as the original argument.  Otherwise, returns a matrix.}
}
\description{
Pads an object to a desired length, either with replicates of itself or another repeated object.
}
\examples{
buffer(c(1,2,3),20)
buffer(matrix(c(1,2,3,4),nrow=2),20)
buffer(list(1,2,3),20)
}

