% Generated by roxygen2 (4.0.0): do not edit by hand
\name{insertRows}
\alias{insertRows}
\title{Inserts a matrix into another matrix.}
\usage{
insertRows(existing, insert, r)
}
\arguments{
\item{existing}{table to insert into}

\item{insert}{rows to insert}

\item{r}{index at which to insert}
}
\description{
Inserts a matrix or data frame into another matrix or data frame.  The new
rows are placed together at the row index specified.
}
\examples{
df1<-data.frame(a=c(1,2,3),b=c(1,2,3),c=c(1,2,3))
insertRows(df1,data.frame(list('a','a','a')),5)
insertRows(df1,data.frame(list('a','a','a')),4)
insertRows(df1,data.frame(list('a','a','a')),3)
insertRows(df1,data.frame(list('a','a','a')),2)
insertRows(df1,data.frame(list('a','a','a')),1)
insertRows(df1,df1,3)
}

