% Generated by roxygen2 (4.0.0): do not edit by hand
\name{len}
\alias{len}
\title{Allows finding the 'length' without knowledge of dimensionality.}
\usage{
len(data)
}
\arguments{
\item{data}{any \code{R} object}
}
\description{
Allows finding the 'length' without knowledge of dimensionality.
}
\examples{
len(list(1,2,3))
len(c(1,2,3,4))
df<-data.frame(a=c(1,2,3),b=c(1,2,3))
len(df)
}

