% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowr.R
\docType{package}
\name{rowr}
\alias{rowr}
\alias{rowr-package}
\title{Row-based functions for R objects}
\description{
Rowr allows the manipulation of R objects as if they were organized rows in a
way that is familiar to people used to working with databases.  It allows
more consistent and predictable output to common functions, and generalizes a
number of utility functions to to be failsafe with any number of objects.
}

