% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowr.R
\name{rows}
\alias{rows}
\title{Allows row indexing without knowledge of dimensionality or class.}
\usage{
rows(data, rownums)
}
\arguments{
\item{data}{any \code{R} object}

\item{rownums}{indices of target rows}
}
\description{
Allows row indexing without knowledge of dimensionality or class.
}
\examples{
rows(c('A','B','C'),c(1,3))
rows(list('A','B','C'),c(1,3))
df<-data.frame(a=c(1,2,3),b=c(1,2,3))
rows(df,3)
}

