% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowApply}
\alias{rowApply}
\title{Applies a function row-wise on any data object.}
\usage{
rowApply(data, fun, ...)
}
\arguments{
\item{data}{any \code{R} object}

\item{fun}{the function to evaluate}

\item{...}{additional arguments to pass to \code{fun}
rowApply(list(1,2,3),function (x) sum(unlist(x)))
rowApply(df,sum)}
}
\description{
Essentially functions as a \code{MARGIN=1} \code{\link{apply}} apply but also works on data objects without 2 dimensions such as lists and vectors.
}

