% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowr.R
\docType{package}
\name{rowr}
\alias{rowr}
\alias{rowr-package}
\title{Row-Based Functions for R Objects}
\description{
Provides utilities which interact with all R objects as
if they were arranged in rows.  It allows more consistent and predictable 
output to common functions, and generalizes a number of utility functions to
to be failsafe with any number and type of input objects.
}

